requirements.txt
fastapi==0.116.1
httpx==0.28.1
python-jose==3.5.0
pydantic==2.11.7
gunicorn==22.0.0
uvicorn==0.35.0
aiomysql==0.2.0
PyMySQL==1.1.1
SQLAlchemy==2.0.42
